new systems and components
new physics system, rigid body component, input component.
Basic functionality for handling input and state using the components architecture

+++ src/render_system.cpp
@@ -7,8 +7,8 @@ RenderSystem::RenderSystem()
 
 void RenderSystem::Render(TransformComponent &transformComponent, SpriteComponent &spriteComponent, std::vector<ViewPort> &viewports)
 {
-    const SDL_Rect destinationRectangle = { static_cast<int>(round(transformComponent.Position.x - viewports[0].camera.frame.x)),
-                                            static_cast<int>(round(transformComponent.Position.y - viewports[0].camera.frame.y)),
+    const SDL_Rect destinationRectangle = { static_cast<int>(round(transformComponent.Position.x - static_cast<double>(viewports[0].camera.frame.x))),
+                                            static_cast<int>(round(transformComponent.Position.y - static_cast<double>(viewports[0].camera.frame.y))),
                                             spriteComponent.rect.w,
                                             spriteComponent.rect.h};
